VERY Basic Attack (99% Complete)
Spawns in invisible box sprite aka the hit-box, and then deletes it immediately. Just need to add a timer so it lasts X amt of time for whatever the duration of the actual attack will last then it will be done. Also added mouse input for both the hook and this attack in this branch but will merge it with the main one later.